adjust angular-promise-tracker typings to types-2.0
